requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pip==21.1.3
telethon==1.16.4
python-dotenv==0.18.0
setuptools==57.4.0
Werkzeug==2.0.1
asgiref==3.4.1
python-decouple